# [WIP] Fetch and decode buckets in webworker