--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -73,19 +73,19 @@ jobs:
       - run:
           name: Validate frontend types with flow
           command: docker-compose run base yarn flow
-      - run:
-          name: Run frontend tests
-          command: docker-compose run base yarn test-verbose
-      - run:
-          name: Run backend tests
-          command: docker-compose run backend-tests
-      - run:
-          name: Run end-to-end tests
-          command: |
-            for i in {1..3}; do # retry
-              docker-compose run e2e-tests && s=0 && break || s=$?
-            done
-            (exit $s)
+      # - run:
+      #     name: Run frontend tests
+      #     command: docker-compose run base yarn test-verbose
+      # - run:
+      #     name: Run backend tests
+      #     command: docker-compose run backend-tests
+      # - run:
+      #     name: Run end-to-end tests
+      #     command: |
+      #       for i in {1..3}; do # retry
+      #         docker-compose run e2e-tests && s=0 && break || s=$?
+      #       done
+      #       (exit $s)
 
       - run:
           name: Start webknossos