--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -145,9 +145,9 @@ jobs:
             fi
             docker logout
 
-      - run:
-          name: Report coverage
-          command: docker-compose run base yarn coverage
+      # - run:
+      #     name: Report coverage
+      #     command: docker-compose run base yarn coverage
 workflows:
   version: 2
   default: