Even though, async fetching of data is [supposed](https://stackoverflow.com/questions/18768452/web-workers-handling-ajax-calls-optimisation-overkill/) to not block the main thread, I noticed performance improvements on my setup (speed up factors reaching from 1.3 to above 2.0).

The code in this PR is hacked together as a prototype. If we decide to go with webworkers, we should think about how we want to structure the code so that code duplication is avoided.

My setup was:
- load a dataset with segmentation so that zoom step 0 is used (shortly before zoom step 1 to maximize the bucket count)
- set segmentation opacity to zero and refresh the page
- wait until after loading is completed
- [optional: set window.useWebWorker = true to test with web worker]
- start the perf measurement
- change seg opacity from 0 to 50 (via keyboard; not slider)
- after all segmentation buckets are loaded, use chrome's screenshot timeline to see the exact interval in which buckets were loaded (i.e., the first screenshot showing 50, and the first screenshot showing complete data)

The overall time did reduce drastically for me, when the webworker was active (JS execution time itself didn't change very much). Also, the idle time was higher (above 100ms instead of 1 ms).

We should try to reproduce these findings on at least one other system.

In general, using a webworker introduces more complexity (e.g., due to the code isolation). Also it's not clear to me why this perf improvement exists. It would make sense, if the perf speed up was due to the decoding of 4 bit data, but segmentation data is never encoded in 4-bit...

### URL of deployed dev instance (used for testing):
- https://bucketwebworker.webknossos.xyz

### Steps to test:
- abc

### Issues:
- fixes #2967 

------
- [ ] Updated [changelog](../blob/master/CHANGELOG.md#unreleased)
- [ ] Updated [migration guide](../blob/master/MIGRATIONS.md#unreleased) if applicable
- [ ] Needs datastore update after deployment
- [ ] Ready for review