@daniel-wer 2018-08-20
Very cool :)

I'll add my measured times (measured on https://bucketwebworker.webknossos.xyz). In contrast to you, I measured the time from changing the segmentation opacity to 50 until ALL buckets were loaded (as it was not easily recognizable for me when the viewports were fully loaded from the screenshots, so I used the network display):

Without webworker: 8.18s (100ms Idle)
With webworker: 6.38s (2047ms Idle)

Very impressive, especially the huge difference in idle time!
I also don't know where exactly this speedup is coming from, the only thing I noticed is that without the webworker, Chrome spends more than half of the execution time doing "Other" stuff (not sure what this is). With a webworker this time spent doing "Other" stuff is decreased considerably.