@philippotto 2018-09-06
Closing this in favor of #3162, which implements the webworker approach properly.